el tipo de dato de telefono, de la clase granja fue cambiado de int a long

+++ src/main/java/com/sistema/inventario/models/Granja.java
@@ -21,7 +21,7 @@ public class Granja {
 	private String nombre;
 	
 	@Column(name="telefono", length=50, nullable=false)	
-	private int telefono;
+	private long telefono;
 	
 	@Column(name="direccion", length=50, nullable=false)	
 	private String direccion;
@@ -72,7 +72,7 @@ public class Granja {
 
 
 
-	public Granja(Long id, String nombre, int telefono, String direccion, double dinero, int capacidadHuevos,
+	public Granja(Long id, String nombre, long telefono, String direccion, double dinero, int capacidadHuevos,
 			int capacidadPollos, double precioVentaPollo, double precioVentaHuevo, double precioCompraPollo,
 			double precioCompraHuevo, int diasDeCompraPollo, int diasDeCompraHuevo) {
 		super();
@@ -100,7 +100,7 @@ public class Granja {
 
 
 
-	public Granja(String nombre, int telefono, String direccion, double dinero, int capacidadHuevos,
+	public Granja(String nombre, long telefono, String direccion, double dinero, int capacidadHuevos,
 			int capacidadPollos, double precioVentaPollo, double precioVentaHuevo, double precioCompraPollo,
 			double precioCompraHuevo, int diasDeCompraPollo, int diasDeCompraHuevo) {
 		super();
@@ -237,7 +237,7 @@ public class Granja {
 
 
 
-	public int getTelefono() {
+	public long getTelefono() {
 		return telefono;
 	}
 
@@ -247,7 +247,7 @@ public class Granja {
 
 
 
-	public void setTelefono(int telefono) {
+	public void setTelefono(long telefono) {
 		this.telefono = telefono;
 	}
